Fix edge cases with Lint/UnusedComparison and Lint/UnusedLiteral
Comparisons with a regex using `===` and `=~` can have side effects, so they should not be counted.
If the right of an And/Or is a call, the left should be counted, as it potentially has side effects / will control whether the right is called.
If a call has args, count its object as being used, as it could have side effects.
Fix yield and literals in generics having their expressions counted as unused.